[rt|ethernet] drivers/net: fix livelock issues
Preempt-RT runs into a live lock issue with the NETDEV_TX_LOCKED micro
optimization. The reason is that the softirq thread is rescheduling
itself on that return value. Depending on priorities it starts to
monoplize the CPU and livelock on UP systems.

Remove it.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>

(cherry picked from commit 7e9e5e9152dc38515606032d338664cc119b7043)
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>