[rt] i2c/omap: drop the lock hard irq context
The lock is taken while reading two registers. On RT the first lock is
taken in hard irq where it might sleep and in the threaded irq.
The threaded irq runs in oneshot mode so the hard irq does not run until
the thread the completes so there is no reason to grab the lock.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
(cherry picked from commit a05139082c2e70519899a17d032c57fe2777999a)
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>