[rt] kernel/hotplug: restore original cpu mask oncpu/down
If a task which is allowed to run only on CPU X puts CPU Y down then it
will be allowed on all CPUs but the on CPU Y after it comes back from
kernel. This patch ensures that we don't lose the initial setting unless
the CPU the task is running is going down.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
(cherry picked from commit 3e4cb0daa2873cbe267c5ef2378d91fd4a38a5b6)
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>