clk: msm: mdss: add support for chip serial number in dfps info
Dfps info is calculated for a specific msm part and needs to be
used for the same. Add a chip serial number for the dfps info so
that the PLL trim codes are used for the same msm part on which
they are generated.

Change-Id: I6f1e519c1e9278b35dfbdf3d9ba4a2ed0e1a6a78
Signed-off-by: Ashish Garg <ashigarg@codeaurora.org>
(cherry picked from commit 32d80f8c0cee35df7f06fdc3209bd6e86d94e5db)
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>