msm: kgsl: Don't read stale crashdumper data
After the crashdumper runs once, and the user sets the
snapshot_crashdumper sysfs node to 0, it is possible that stale data
is saved off into the snapshot. Check the value of the sysfs node at
the appropriate place to force a re-read of the data through the
legacy path.

Change-Id: Icf27c39b68d786003c33fd0eb22bd3e8b9a036be
Signed-off-by: Lynus Vaz <lvaz@codeaurora.org>
(cherry picked from commit 1e450489ff0d95b68f14c6a0ca0b392334ea2653)
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>