msm: kgsl: Look for freed entries with the pagetable name
The list of freed entries stores the pagetable name which is usually
the process id. Search through this list using the process id and not
the thread id so that we see the entry we're interested in.

Change-Id: I4b19d8f47be0973b4f1ecf7dc2641b1824c6ca56
Signed-off-by: Lynus Vaz <lvaz@codeaurora.org>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
(cherry picked from commit 2e056c800541ebfd596ed8e59ee761d15ac43c02)
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>